fix<Accounting>
Implementasi fitur maintenance nota penjualan tunai
Implementasi Update Surat Jalan untuk Jual Tunai
bug fix pada fitur penagihan penjualan untuk perhitungan total pelunasan

+++ public/js/Accounting/Piutang/PenjualanLokal/PenagihanPenjualan.js
@@ -546,7 +546,7 @@ $(document).ready(function () {
             .toArray()
             .reduce((sum, row) => {
                 let jumlahUang = numeral(row[4]).value();
-                return sum + parseInt(jumlahUang);
+                return sum + parseFloat(jumlahUang);
             }, 0);
 
         let dppNilaiLain = (totalPelunasan * 11) / 12;
